refactor: modularize mobile server architecture and fix path issues
- Split monolithic mobile server into modular components (auth, routes, utils, websocket)
- Fix static file path resolution after directory restructure
- Add URL decoding for file content API to handle encoded paths
- Improve code organization and maintainability
