+++ src/services/mobile/auth/index.ts
@@ -1,0 +1,1 @@
+export { AuthManager, AuthConfig } from './auth-manager';
