+++ src/services/mobile/routes/index.ts
@@ -1,0 +1,2 @@
+export { APIRoutes } from './api-routes';
+export { StaticRoutes } from './static-routes';
