+++ src/services/mobile/search/index.ts
@@ -1,0 +1,2 @@
+export { FileSearchService } from './file-search';
+export { FileExplorer } from './file-explorer';
