+++ src/services/mobile/types/index.ts
@@ -1,0 +1,1 @@
+export { MESSAGE_TYPES } from './message-types';
