+++ src/services/mobile/types/message-types.ts
@@ -1,0 +1,7 @@
+
+export const MESSAGE_TYPES = {
+  INITIAL_STATE: 'initialState',
+  QUEUE_UPDATE: 'queueUpdate',
+  STATUS_UPDATE: 'statusUpdate',
+  OUTPUT_UPDATE: 'outputUpdate',
+} as const;
