+++ src/services/mobile/utils/index.ts
@@ -1,0 +1,1 @@
+export { FileUtils } from './file-utils';
