+++ src/services/mobile/websocket/index.ts
@@ -1,0 +1,1 @@
+export { WebSocketManager } from './websocket-manager';
